requirements.txt
Django==1.10.6
argparse==1.2.1
django-debug-toolbar==1.7
django-extensions==1.7.8
django-haystack==2.6.0
djangorestframework==3.6.2
drf-haystack==1.6.1
elasticsearch==5.3.0
loremipsum==1.0.5
psycopg2==2.7.1
python-dateutil==2.6.0
six==1.10.0
sqlparse==0.2.3
urllib3==1.20
wsgiref==0.1.2
